fix: SaveChatRequestDto roomType 타입수정

+++ campton/campton/src/main/java/com/dongguk/campton/dto/request/SaveChatRequestDto.java
@@ -14,7 +14,7 @@ public class SaveChatRequestDto {
     //	"answer": "String",
     //	"question": "String"
     private Long memberId;
-    private Long roomType;
+    private Integer roomType;
     private String answer;
     private String question;
